requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clingo==5.7.1
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dafsa==1.0
data==0.4
debugpy==1.8.0
decorator==4.4.2
dm-tree==0.1.8
docstring_parser==0.16
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
funcsigs==1.0.2
future==1.0.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.63.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
hydra-core==1.3.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
# install==1.3.5
ipykernel==6.29.5
ipython==8.25.0
jax-jumpy==1.0.0
jedi==0.19.1
joblib==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
latex==0.7.0
lazy_loader==0.4
lz4==4.3.3
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
minigrid==2.3.1
moviepy==1.0.3
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
optuna==3.6.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-sat==0.1.7.dev26
pytictoc==1.5.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.24.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
# sb3_contrib==2.4.0
scikit-image==0.23.2
scipy==1.13.0
shellingham==1.5.4
shtab==1.7.1
shutilwhich==1.1.0
six==1.16.0
SQLAlchemy==2.0.30
# stable_baselines3==2.4.1
stack-data==0.6.3
tempdir==0.7.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.2.0
tifffile==2024.5.10
torch==1.13.1
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
